Align with Q-Marth's change to the servo's type signatures.

Conflicts:
	include/dbg/servo.c
	include/live/servo.c

+++ include/dbg/servo.c
@@ -4,11 +4,10 @@
 
 void servo_init() {}
 
-int servo_set_pos(int angle){
+uint16_t servo_set_pos(int angle){
     return angle;
 }
 
-int servo_get_pos(){
-    return dequeue(SERVO_QUEUE, uint32_t);
-
+uint16_t servo_get_pos(){
+    return dequeue(SERVO_QUEUE, uint16_t);
 }
